+++ example_site/users/views.py
@@ -1,4 +1,4 @@
-from django.views.generic import DetailView, ListView
+from django.views.generic import DetailView
 
 from django.contrib.auth.mixins import LoginRequiredMixin
 
@@ -11,9 +11,5 @@ class UserDetailView(LoginRequiredMixin, DetailView):
     slug_field = 'username'
     slug_url_kwarg = 'username'
 
-
-class UserListView(LoginRequiredMixin, ListView):
-    model = User
-    # These next two lines tell the view to index lookups by username
-    slug_field = 'username'
-    slug_url_kwarg = 'username'
+    # Projectroles dependency
+    template_name = 'projectroles/user_detail.html'